move NetworkStat to own file
